Move compe from lsp/ to dev/
